# identity/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Guess Who",
        "description": "Documentation pour l'API Guess Who.",
        "contact": {
            "email": "support@votre-site.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/token": {
            "get": {
                "tags": [
                    "token"
                ],
                "summary": "Génère un nouveau token",
                "description": "Génère un nouveau token pour l'application.",
                "operationId": "5ac16309893cb565f49fd1b30848d06a",
                "responses": {
                    "200": {
                        "description": "Token généré avec succès.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "type": "string",
                                            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erreur serveur.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Erreur interne."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/token/generate/{id_account}": {
            "get": {
                "tags": [
                    "token"
                ],
                "summary": "Génère un modèle de token pour un utilisateur",
                "description": "Génère un modèle de token pour un utilisateur spécifique en fonction de l'ID de compte.",
                "operationId": "78cc5e1bebc825f13bd2abdae48bcef8",
                "parameters": [
                    {
                        "name": "id_account",
                        "in": "path",
                        "description": "ID de l'utilisateur pour générer le modèle de token.",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 123
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Modèle de token généré avec succès.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token_model": {
                                            "type": "string",
                                            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "399": {
                        "description": "Erreur lors de la génération du token.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "ID de compte invalide."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/token/regenerate/{id_account}": {
            "get": {
                "tags": [
                    "token"
                ],
                "summary": "Regénère un token pour un utilisateur",
                "description": "Regénère un token pour un utilisateur spécifique en fonction de l'ID de compte.",
                "operationId": "b7a54a6d3d44cafea979219e47a28de2",
                "parameters": [
                    {
                        "name": "id_account",
                        "in": "path",
                        "description": "ID de l'utilisateur pour régénérer le token.",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 123
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Token régénéré avec succès.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "type": "string",
                                            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "399": {
                        "description": "Erreur lors de la régénération du token.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "ID de compte invalide."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
...
}